SCB-85 move write file from example to extension
