+++ metrics/pom.xml
@@ -31,8 +31,9 @@
   <artifactId>metrics</artifactId>
 
   <modules>
-    <module>metrics-core</module>
-    <module>metrics-sample</module>
     <module>metrics-common</module>
+    <module>metrics-core</module>
+    <module>metrics-extension</module>
+    <module>metrics-sample</module>
   </modules>
 </project>
